don't modify tree.args while evaluating a function
jme.evaluate evaluated args of an op or function in-place while
evaluating the function. This meant that the input tree was modified
as a side effect of evaluating it. That's bad!

Now, a new list of args in local scope is created to hold the results, and the input
tree is not modified.